+++ src/app/dashboard/forms/estilosFormularios.css
@@ -10,7 +10,7 @@
 }
 
 body {
-    background: url("/img/ImgForm.jpg");
+    background: url("/img/imgform.jpg");
     background-size: cover;
     background-repeat: no-repeat;
     background-attachment: fixed;
