+++ src/app/dashboard/forms/office/page.tsx
@@ -1,5 +1,5 @@
 import { CiUser } from 'react-icons/ci';
-import '../EstilosFormularios.css';
+import '../estilosFormularios.css';
 
 function fechaActual() {
   let fecha = new Date();
@@ -14,35 +14,40 @@ function obtenerId_Ciuddad() {
   // Obtener el id de la ciudad usando Post en la ruta /api/ciudad
 }
 
-export default function formClients() {
+export default function formOffice() {
   return (
     <div className='contenedor'>
       <div className='caja-formulario'>
         <div className='contenedor-login' id='login'>
           <div className='encabezado'>
-            <h1 className='titulo'>Ingresar Cliente</h1>
+            <h1 className='titulo'>Registro de Sucursal</h1>
           </div>
           <div className='caja-entrada'>
             <input
               type='text'
               className='entrada'
-              placeholder='Nombre del cliente'
+              placeholder='Nombre de la sucursal'
             />
             <i className='logo'>
               <CiUser />
             </i>
           </div>
           <div className='caja-entrada'>
-            <input type='text' className='entrada' placeholder='Ciudad' />
+            <input
+              type='text'
+              className='entrada'
+              placeholder='Nombre de la ciudad'
+            />
             <i className='logo'>
               <CiUser />
             </i>
           </div>
+
           <div className='caja-entrada'>
             <input
               type='submit'
               className='boton-enviar'
-              value='Crear Cliente'
+              value='Registrar Sucursal'
             />
           </div>
         </div>
